packages/scraper/pyproject.toml
[project]
name = "lumina-forge-scraper"
version = "0.1.0"
description = "Scrapes Expedition 33 guides and writes structured data into the Obsidian vault."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Léon Fievet" }]

dependencies = [
  "httpx>=0.28",
  "beautifulsoup4>=4.12",
  "lxml>=5.3",
  "pydantic>=2.10",
  "python-frontmatter>=1.1",
  "rich>=13.9",
  "tenacity>=9.0",
  "typer>=0.15",
]

[project.scripts]
scraper = "scraper.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scraper"]
